chore: move pipelines types to own file
