+++ packages/playground/src/types.ts
@@ -10,6 +10,7 @@ export * from './api/documents/shared';
 export * from './api/documents/classifierTypes';
 export * from './api/documents/feedbackTypes';
 export * from './api/documents/previewTypes';
+export * from './api/documents/pipelinesTypes';
 
 // This file is here mostly to allow apis to import { ... } from '../../types';
 // Overriding types should probably be done in their respective API endpoint files, where possible
